requirements.txt
tesseract
python-telegram-bot
Image
Pillow
telegram == 0.0.1